Upgraded HandGameplay and plugins to Unreal Engine 5.2
- Switched Oculus plugin references to new versions.
- Removed `WITH_OCULUS_ENGINE` defines, since `HandTrackingFilter` no longer requires the Oculus fork.

+++ Plugins/OculusHandTools/Source/OculusInteractable/Private/TransformString.h
@@ -33,7 +33,7 @@ private:
 	 * @param Destination - Output float.
 	 * @return bool - True if prefix matches.
 	 */
-	static bool ReadFloat(FString& Source, const TCHAR* ExpectedPrefix, float& Destination)
+	static bool ReadFloat(FString& Source, const TCHAR* ExpectedPrefix, double& Destination)
 	{
 		if (**Source != *ExpectedPrefix) return false;
 		Destination = FCString::Atof(*Source + 1);
